+++ src/main/java/org/plumelib/util/ClassDeterministic.java
@@ -74,6 +74,10 @@ public class ClassDeterministic {
    * @param c the Class whose enum constants to return
    * @return the class's enum constants
    */
+  @SuppressWarnings({
+    "signedness", // ToStringComparator problem
+    "mustcall:argument" // not sure; generics inference problem?
+  })
   public static <T> T @Nullable [] getEnumConstants(Class<T> c) {
     @NonNull T[] result = c.getEnumConstants();
     if (result == null) {
@@ -178,6 +182,9 @@ public class ClassDeterministic {
   /** Compares Annotation objects by type name. */
   private static class AnnotationComparator implements Comparator<Annotation> {
 
+    /** Create a new AnnotationComparator. */
+    public AnnotationComparator() {}
+
     @Override
     public int compare(Annotation a1, Annotation a2) {
       return classComparator.compare(a1.annotationType(), a2.annotationType());
@@ -189,6 +196,9 @@ public class ClassDeterministic {
 
   /** Compares Class objects by fully-qualified name. */
   private static class ClassComparator implements Comparator<Class<?>> {
+
+    /** Create a new ClassComparator. */
+    public ClassComparator() {}
 
     @Override
     public int compare(Class<?> c1, Class<?> c2) {
@@ -208,6 +218,9 @@ public class ClassDeterministic {
    */
   private static class MethodComparator implements Comparator<Method> {
 
+    /** Create a new MethodComparator. */
+    public MethodComparator() {}
+
     @Override
     public int compare(Method m1, Method m2) {
       int result;
@@ -223,7 +236,7 @@ public class ClassDeterministic {
       }
       assert ptypes1.length == ptypes2.length
           : "@AssumeAssertion(index): difference of lengths is 0;"
-                + " https://github.com/kelloggm/checker-framework/issues/231";
+              + " https://github.com/kelloggm/checker-framework/issues/231";
       for (int i = 0; i < ptypes1.length; i++) {
         result = classComparator.compare(ptypes1[i], ptypes2[i]);
         if (result != 0) {
@@ -258,6 +271,9 @@ public class ClassDeterministic {
    */
   private static class ConstructorComparator implements Comparator<Constructor<?>> {
 
+    /** Create a new ConstructorComparator. */
+    public ConstructorComparator() {}
+
     @Override
     public int compare(Constructor<?> c1, Constructor<?> c2) {
       int result = classComparator.compare(c1.getDeclaringClass(), c2.getDeclaringClass());
@@ -272,7 +288,7 @@ public class ClassDeterministic {
       }
       assert ptypes1.length == ptypes2.length
           : "@AssumeAssertion(index): difference of lengths is 0;"
-                + " https://github.com/kelloggm/checker-framework/issues/231";
+              + " https://github.com/kelloggm/checker-framework/issues/231";
       for (int i = 0; i < ptypes1.length; i++) {
         result = classComparator.compare(ptypes1[i], ptypes2[i]);
         if (result != 0) {
@@ -289,6 +305,9 @@ public class ClassDeterministic {
   /** Compares Field objects by name. */
   private static class FieldComparator implements Comparator<Field> {
 
+    /** Create a new FieldComparator. */
+    public FieldComparator() {}
+
     @Override
     public int compare(Field f1, Field f2) {
       int result = classComparator.compare(f1.getDeclaringClass(), f2.getDeclaringClass());
@@ -304,6 +323,9 @@ public class ClassDeterministic {
 
   /** Compares objects by the result of toString(). */
   private static class ToStringComparator implements Comparator<Object> {
+
+    /** Create a new ToStringComparator. */
+    public ToStringComparator() {}
 
     @Override
     public int compare(Object o1, Object o2) {