+++ src/main/java/org/plumelib/util/MathPlume.java
@@ -12,6 +12,7 @@ import org.checkerframework.checker.index.qual.PolyUpperBound;
 import org.checkerframework.checker.lock.qual.GuardSatisfied;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import org.checkerframework.checker.signedness.qual.Unsigned;
 import org.checkerframework.common.value.qual.ArrayLen;
 import org.checkerframework.common.value.qual.MinLen;
 import org.checkerframework.common.value.qual.StaticallyExecutable;
@@ -255,7 +256,7 @@ public final class MathPlume {
    */
   @Pure
   @StaticallyExecutable
-  public static int rshiftUnsigned(int x, int y) {
+  public static @Unsigned int rshiftUnsigned(@Unsigned int x, int y) {
     return x >>> y;
   }
 
@@ -268,7 +269,7 @@ public final class MathPlume {
    */
   @Pure
   @StaticallyExecutable
-  public static long rshiftUnsigned(long x, long y) {
+  public static @Unsigned long rshiftUnsigned(@Unsigned long x, long y) {
     return x >>> y;
   }
 
@@ -1482,7 +1483,7 @@ public final class MathPlume {
    *
    * @param missing the missing integers; modified by this method
    * @return value to be returned by {@link #nonmodulusStrict(int[])}: a tuple of (r,m) where all
-   *     numbers in {@code missing} are equal to r (ood m)
+   *     numbers in {@code missing} are equal to r (mod m)
    */
   private static int @Nullable @ArrayLen(2) [] nonmodulusStrictIntInternal(
       Iterator<Integer> missing) {
@@ -1503,6 +1504,8 @@ public final class MathPlume {
   }
 
   /**
+   * Return true if the first and last elements are not equal to r (mod m).
+   *
    * @param rm a tuple of (r,m)
    * @param rfali a sequence of numbers, plus a first and last element outside their range. This
    *     iterator has already been iterated all the way to its end.
@@ -1834,7 +1837,9 @@ public final class MathPlume {
   }
 
   /**
-   * @param rm an an array containing two elements
+   * Return true if the first and last elements are equal to r (mod m).
+   *
+   * @param rm an array containing two elements
    * @param rfali a sequence of numbers, plus a first and last element outside their range. This
    *     iterator has already been iterated all the way to its end.
    * @return true if the first and last elements are equal to r (mod m)