+++ src/main/java/org/plumelib/util/UtilPlume.java
@@ -3,36 +3,27 @@
 package org.plumelib.util;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static java.nio.file.StandardOpenOption.APPEND;
-import static java.nio.file.StandardOpenOption.CREATE;
 
-import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
 import java.io.FilenameFilter;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.LineNumberReader;
 import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
 import java.io.OutputStream;
-import java.io.OutputStreamWriter;
 import java.io.PrintWriter;
 import java.io.Reader;
 import java.io.Serializable;
 import java.io.StringWriter;
-import java.io.Writer;
+import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.BitSet;
@@ -45,14 +36,14 @@ import java.util.Properties;
 import java.util.StringTokenizer;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
-import java.util.zip.GZIPInputStream;
-import java.util.zip.GZIPOutputStream;
 import org.checkerframework.checker.index.qual.IndexOrHigh;
 import org.checkerframework.checker.index.qual.LTEqLengthOf;
 import org.checkerframework.checker.index.qual.NonNegative;
 import org.checkerframework.checker.nullness.qual.NonNull;
 import org.checkerframework.checker.nullness.qual.Nullable;
+import org.checkerframework.checker.nullness.qual.PolyNull;
 import org.checkerframework.checker.regex.qual.Regex;
+import org.checkerframework.checker.signedness.qual.Signed;
 import org.checkerframework.common.value.qual.StaticallyExecutable;
 import org.checkerframework.dataflow.qual.Pure;
 import org.checkerframework.dataflow.qual.SideEffectFree;
@@ -203,17 +194,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2020-02-20
   public static InputStream fileInputStream(Path path) throws IOException {
-    InputStream in;
-    if (path.endsWith(".gz")) {
-      try {
-        in = new GZIPInputStream(new FileInputStream(path.toFile()));
-      } catch (IOException e) {
-        throw new IOException("Problem while reading " + path, e);
-      }
-    } else {
-      in = new FileInputStream(path.toFile());
-    }
-    return in;
+    return FilesPlume.newFileInputStream(path);
   }
 
   /**
@@ -231,17 +212,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static InputStream fileInputStream(File file) throws IOException {
-    InputStream in;
-    if (file.getName().endsWith(".gz")) {
-      try {
-        in = new GZIPInputStream(new FileInputStream(file));
-      } catch (IOException e) {
-        throw new IOException("Problem while reading " + file, e);
-      }
-    } else {
-      in = new FileInputStream(file);
-    }
-    return in;
+    return FilesPlume.newFileInputStream(file);
   }
 
   /**
@@ -261,8 +232,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static InputStreamReader fileReader(String filename)
       throws FileNotFoundException, IOException {
-    // return fileReader(filename, "ISO-8859-1");
-    return fileReader(new File(filename), null);
+    return FilesPlume.newFileReader(filename);
   }
 
   /**
@@ -281,7 +251,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static InputStreamReader fileReader(Path path) throws FileNotFoundException, IOException {
-    return fileReader(path.toFile(), null);
+    return FilesPlume.newFileReader(path);
   }
 
   /**
@@ -302,14 +272,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static InputStreamReader fileReader(Path path, @Nullable String charsetName)
       throws FileNotFoundException, IOException {
-    InputStream in = new FileInputStream(path.toFile());
-    InputStreamReader fileReader;
-    if (charsetName == null) {
-      fileReader = new InputStreamReader(in, UTF_8);
-    } else {
-      fileReader = new InputStreamReader(in, charsetName);
-    }
-    return fileReader;
+    return FilesPlume.newFileReader(path, charsetName);
   }
 
   /**
@@ -328,7 +291,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static InputStreamReader fileReader(File file) throws FileNotFoundException, IOException {
-    return fileReader(file, null);
+    return FilesPlume.newFileReader(file);
   }
 
   /**
@@ -349,14 +312,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static InputStreamReader fileReader(File file, @Nullable String charsetName)
       throws FileNotFoundException, IOException {
-    InputStream in = new FileInputStream(file);
-    InputStreamReader fileReader;
-    if (charsetName == null) {
-      fileReader = new InputStreamReader(in, UTF_8);
-    } else {
-      fileReader = new InputStreamReader(in, charsetName);
-    }
-    return fileReader;
+    return FilesPlume.newFileReader(file, charsetName);
   }
 
   /**
@@ -376,7 +332,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static BufferedReader bufferedFileReader(String filename)
       throws FileNotFoundException, IOException {
-    return bufferedFileReader(new File(filename));
+    return FilesPlume.newBufferedFileReader(filename);
   }
 
   /**
@@ -396,7 +352,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static BufferedReader bufferedFileReader(File file)
       throws FileNotFoundException, IOException {
-    return bufferedFileReader(file, null);
+    return FilesPlume.newBufferedFileReader(file);
   }
 
   /**
@@ -417,7 +373,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static BufferedReader bufferedFileReader(String filename, @Nullable String charsetName)
       throws FileNotFoundException, IOException {
-    return bufferedFileReader(new File(filename), charsetName);
+    return FilesPlume.newBufferedFileReader(filename);
   }
 
   /**
@@ -438,8 +394,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static BufferedReader bufferedFileReader(File file, @Nullable String charsetName)
       throws FileNotFoundException, IOException {
-    Reader fileReader = fileReader(file, charsetName);
-    return new BufferedReader(fileReader);
+    return FilesPlume.newBufferedFileReader(file, charsetName);
   }
 
   /**
@@ -459,7 +414,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static LineNumberReader lineNumberFileReader(String filename)
       throws FileNotFoundException, IOException {
-    return lineNumberFileReader(new File(filename));
+    return FilesPlume.newLineNumberFileReader(filename);
   }
 
   /**
@@ -479,18 +434,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static LineNumberReader lineNumberFileReader(File file)
       throws FileNotFoundException, IOException {
-    Reader fileReader;
-    if (file.getName().endsWith(".gz")) {
-      try {
-        fileReader =
-            new InputStreamReader(new GZIPInputStream(new FileInputStream(file)), "ISO-8859-1");
-      } catch (IOException e) {
-        throw new IOException("Problem while reading " + file, e);
-      }
-    } else {
-      fileReader = new InputStreamReader(new FileInputStream(file), "ISO-8859-1");
-    }
-    return new LineNumberReader(fileReader);
+    return FilesPlume.newLineNumberFileReader(file);
   }
 
   /**
@@ -508,7 +452,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static BufferedWriter bufferedFileWriter(String filename) throws IOException {
-    return bufferedFileWriter(filename, false);
+    return FilesPlume.newBufferedFileWriter(filename);
   }
 
   /**
@@ -527,19 +471,9 @@ public final class UtilPlume {
    * @deprecated use {@link FilesPlume#newBufferedFileWriter}
    */
   @Deprecated // deprecated 2021-02-25
-  // Question:  should this be rewritten as a wrapper around bufferedFileOutputStream?
   public static BufferedWriter bufferedFileWriter(String filename, boolean append)
       throws IOException {
-    if (filename.endsWith(".gz")) {
-      return new BufferedWriter(
-          new OutputStreamWriter(
-              new GZIPOutputStream(new FileOutputStream(filename, append)), UTF_8));
-    } else {
-      return Files.newBufferedWriter(
-          Paths.get(filename),
-          UTF_8,
-          append ? new StandardOpenOption[] {CREATE, APPEND} : new StandardOpenOption[] {CREATE});
-    }
+    return FilesPlume.newBufferedFileWriter(filename, append);
   }
 
   /**
@@ -560,11 +494,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   public static BufferedOutputStream bufferedFileOutputStream(String filename, boolean append)
       throws IOException {
-    OutputStream os = new FileOutputStream(filename, append);
-    if (filename.endsWith(".gz")) {
-      os = new GZIPOutputStream(os);
-    }
-    return new BufferedOutputStream(os);
+    return FilesPlume.newBufferedFileOutputStream(filename, append);
   }
 
   ///////////////////////////////////////////////////////////////////////////
@@ -946,14 +876,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static void writeObject(Object o, File file) throws IOException {
-    // 8192 is the buffer size in BufferedReader
-    OutputStream bytes = new BufferedOutputStream(new FileOutputStream(file), 8192);
-    if (file.getName().endsWith(".gz")) {
-      bytes = new GZIPOutputStream(bytes);
-    }
-    ObjectOutputStream objs = new ObjectOutputStream(bytes);
-    objs.writeObject(o);
-    objs.close();
+    FilesPlume.writeObject(o, file);
   }
 
   /**
@@ -970,17 +893,7 @@ public final class UtilPlume {
   @Deprecated // deprecated 2021-02-25
   @SuppressWarnings("BanSerializableRead") // wrapper around dangerous API
   public static Object readObject(File file) throws IOException, ClassNotFoundException {
-    // 8192 is the buffer size in BufferedReader
-    InputStream istream = new BufferedInputStream(new FileInputStream(file), 8192);
-    if (file.getName().endsWith(".gz")) {
-      try {
-        istream = new GZIPInputStream(istream);
-      } catch (IOException e) {
-        throw new IOException("Problem while reading " + file, e);
-      }
-    }
-    ObjectInputStream objs = new ObjectInputStream(istream);
-    return objs.readObject();
+    return FilesPlume.readObject(file);
   }
 
   /**
@@ -993,17 +906,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static String readerContents(Reader r) {
-    try {
-      StringBuilder contents = new StringBuilder();
-      int ch;
-      while ((ch = r.read()) != -1) {
-        contents.append((char) ch);
-      }
-      r.close();
-      return contents.toString();
-    } catch (Exception e) {
-      throw new Error("Unexpected error in readerContents(" + r + ")", e);
-    }
+    return FilesPlume.readerContents(r);
   }
 
   // an alternate name would be "fileContents".
@@ -1020,22 +923,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static String readFile(File file) {
-
-    try {
-      BufferedReader reader = UtilPlume.bufferedFileReader(file);
-      StringBuilder contents = new StringBuilder();
-      String line = reader.readLine();
-      while (line != null) {
-        contents.append(line);
-        // Note that this converts line terminators!
-        contents.append(lineSep);
-        line = reader.readLine();
-      }
-      reader.close();
-      return contents.toString();
-    } catch (Exception e) {
-      throw new Error("Unexpected error in readFile(" + file + ")", e);
-    }
+    return FilesPlume.readFile(file);
   }
 
   /**
@@ -1049,14 +937,7 @@ public final class UtilPlume {
    */
   @Deprecated // deprecated 2021-02-25
   public static void writeFile(File file, String contents) {
-
-    try {
-      Writer writer = Files.newBufferedWriter(file.toPath(), UTF_8);
-      writer.write(contents, 0, contents.length());
-      writer.close();
-    } catch (Exception e) {
-      throw new Error("Unexpected error in writeFile(" + file + ")", e);
-    }
+    FilesPlume.writeFile(file, contents);
   }
 
   ///////////////////////////////////////////////////////////////////////////
@@ -1273,7 +1154,8 @@ public final class UtilPlume {
    * @deprecated use {@link StringsPlume#mapToStringAndClass}
    */
   @Deprecated // use StringsPlume.mapToStringAndClass; deprecated 2020-12-02
-  public static String mapToStringAndClass(Map<?, ?> m) {
+  public static String mapToStringAndClass(
+      Map<? extends @Signed @PolyNull Object, ? extends @Signed @PolyNull Object> m) {
     return StringsPlume.mapToStringAndClass(m);
   }
 
@@ -1742,7 +1624,7 @@ public final class UtilPlume {
    */
   @Deprecated // use join(CharSequence, Object...) which has the arguments in the other order;
   // deprecated 2020-02-20
-  public static <T> String join(T[] a, CharSequence delim) {
+  public static <T> String join(@Signed T[] a, CharSequence delim) {
     if (a.length == 0) {
       return "";
     }
@@ -1773,7 +1655,7 @@ public final class UtilPlume {
    */
   @SafeVarargs
   @Deprecated // use StringsPlume.join; deprecated 2020-12-02
-  public static <T> String join(CharSequence delim, T... a) {
+  public static <T> String join(CharSequence delim, @Signed T... a) {
     if (a.length == 0) {
       return "";
     }
@@ -1799,7 +1681,7 @@ public final class UtilPlume {
   @SafeVarargs
   @SuppressWarnings("varargs")
   @Deprecated // use StringsPlume.joinLines; deprecated 2020-12-02
-  public static <T> String joinLines(T... a) {
+  public static <T> String joinLines(@Signed T... a) {
     return join(lineSep, a);
   }
 
@@ -1819,10 +1701,10 @@ public final class UtilPlume {
    */
   @Deprecated // use join(CharSequence, Iterable) which has the arguments in the other order;
   // deprecated 2020-12-02
-  public static String join(Iterable<?> v, CharSequence delim) {
+  public static String join(Iterable<? extends @Signed @PolyNull Object> v, CharSequence delim) {
     StringBuilder sb = new StringBuilder();
     boolean first = true;
-    Iterator<?> itor = v.iterator();
+    Iterator<? extends @Signed @PolyNull Object> itor = v.iterator();
     while (itor.hasNext()) {
       if (first) {
         first = false;
@@ -1848,10 +1730,10 @@ public final class UtilPlume {
    * @deprecated use {@link StringsPlume#join}
    */
   @Deprecated // use StringsPlume.join; deprecated 2020-12-02
-  public static String join(CharSequence delim, Iterable<?> v) {
+  public static String join(CharSequence delim, Iterable<? extends @Signed @PolyNull Object> v) {
     StringBuilder sb = new StringBuilder();
     boolean first = true;
-    Iterator<?> itor = v.iterator();
+    Iterator<? extends @Signed @PolyNull Object> itor = v.iterator();
     while (itor.hasNext()) {
       if (first) {
         first = false;
@@ -1873,7 +1755,7 @@ public final class UtilPlume {
    * @deprecated use {@link StringsPlume#joinLines}
    */
   @Deprecated // use StringsPlume.joinLines; deprecated 2020-12-02
-  public static String joinLines(Iterable<?> v) {
+  public static String joinLines(Iterable<? extends @Signed @PolyNull Object> v) {
     return join(lineSep, v);
   }
 
@@ -2384,11 +2266,14 @@ public final class UtilPlume {
    *
    * @deprecated use {@link StringsPlume.NullableStringComparator}
    */
-  @Deprecated // use StringsPlume.NullableStringComparator; deprecated 2020-12-02
+  @Deprecated // 2020-12-02
   public static class NullableStringComparator
       implements Comparator<@Nullable String>, Serializable {
     /** Unique identifier for serialization. If you add or remove fields, change this number. */
     static final long serialVersionUID = 20150812L;
+
+    /** Create a new NullableStringComparator. */
+    public NullableStringComparator() {}
 
     @Pure
     @Override
@@ -2419,10 +2304,13 @@ public final class UtilPlume {
    *
    * @deprecated use {@link StringsPlume.ObjectComparator}
    */
-  @Deprecated // use StringsPlume.ObjectComparator; deprecated 2020-12-02
+  @Deprecated // 2020-12-02
   public static class ObjectComparator implements Comparator<@Nullable Object>, Serializable {
     /** Unique identifier for serialization. If you add or remove fields, change this number. */
     static final long serialVersionUID = 20170420L;
+
+    /** Create a new ObjectComparator. */
+    public ObjectComparator() {}
 
     @SuppressWarnings({
       "allcheckers:purity.not.deterministic.call",
@@ -2679,7 +2567,7 @@ public final class UtilPlume {
 
   /**
    * Returns a String representation of the backtrace of the given Throwable. To see a backtrace at
-   * the the current location, do {@code backtrace(new Throwable())}.
+   * the current location, do {@code backtrace(new Throwable())}.
    *
    * @param t the Throwable to obtain a backtrace of
    * @return a String representation of the backtrace of the given Throwable
@@ -2692,17 +2580,18 @@ public final class UtilPlume {
 
   /**
    * Returns a String representation of the stack trace (the backtrace) of the given Throwable. For
-   * a stack trace at the the current location, do {@code stackTraceToString(new Throwable())}.
+   * a stack trace at the current location, do {@code stackTraceToString(new Throwable())}.
    *
    * @param t the Throwable to obtain a stack trace of
    * @return a String representation of the stack trace of the given Throwable
    */
   public static String stackTraceToString(Throwable t) {
-    StringWriter sw = new StringWriter();
-    PrintWriter pw = new PrintWriter(sw);
-    t.printStackTrace(pw);
-    pw.close();
-    String result = sw.toString();
-    return result;
+    try (StringWriter sw = new StringWriter();
+        PrintWriter pw = new PrintWriter(sw)) {
+      t.printStackTrace(pw);
+      return sw.toString();
+    } catch (IOException e) {
+      throw new UncheckedIOException(e);
+    }
   }
 }